+++ app/src/main/java/com/timife/cowryconverter/presentation/viewmodels/data/RateUiModel.kt
@@ -9,5 +9,5 @@ package com.timife.cowryconverter.presentation.viewmodels.data
 data class RateUiModel(
     val symbol: String = "",
     val name: String = "",
-    val rate: Long = 0
+    val rate: Double = 0.0
 )
